pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hera-r"
version = "0.1.0"
description = "Hyper-Evolving Robust Architecture - Research Edition: Online self-evolving LLM framework"
readme = "README.md"
authors = [
    {name = "Chen Leiyang", email = "chleya@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["llm", "evolution", "online-learning", "transformer", "research"]
dependencies = [
    "torch>=2.1.0",
    "transformer-lens>=1.14.0",
    "sae-lens>=1.3.0",
    "einops>=0.7.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "accelerate>=0.26.0",
    "gradio>=4.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.22.0",
]

[project.urls]
Homepage = "https://github.com/chleya/hera-r"
Documentation = "https://chleya.github.io/hera-r/"
Repository = "https://github.com/chleya/hera-r.git"
Issues = "https://github.com/chleya/hera-r/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["hera*"]
exclude = ["tests*", "examples*", "docs*", "scripts*"]

[tool.setuptools.package-data]
"hera" = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[project.scripts]
hera-cli = "hera.cli:main"
hera-web = "hera.web:main"